# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanban API",
    "version": "v1",
    "summary": "Read and change a board from a script.",
    "description": "Hand-written, not generated. Two tests hold it to the code: one fails when a route is registered and missing here or the other way round, one validates every handler response in the test suite against the schema below.\n\nThere are no tokens. The API grants what the board grants, which is everything to whoever can reach the port, so a deployment puts the same proxy or Cloudflare Access in front of it. Where identity is configured it is used for two things: \"@me\" as an assignee, and the author of a comment.\n\nA POST or PUT body must be application/json. Unknown fields are refused with a 400 that names the field, because \"titel\" for \"title\" would otherwise store a card with no title and nobody would find out until they read the board."
  },
  "servers": [
    {
      "url": "/",
      "description": "The board itself. The paths below are absolute, so they work against any host that serves it."
    }
  ],
...
  "paths": {
...
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "openapi",
        "tags": [
          "meta"
        ],
        "summary": "This document.",
        "description": "Served from the binary, so the spec a deployment answers with is the spec that deployment was built from.",
        "responses": {
          "200": {
            "description": "The OpenAPI document.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards": {
      "get": {
        "operationId": "listBoards",
        "tags": [
          "boards"
        ],
        "summary": "Every board.",
        "responses": {
          "200": {
            "description": "The boards, each with its columns and labels.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createBoard",
        "tags": [
          "boards"
        ],
        "summary": "Create a board.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The board, with the columns it was created with.",
            "headers": {
              "Location": {
                "description": "Where the new board can be read.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "A board with that slug exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "get": {
        "operationId": "getBoard",
        "tags": [
          "boards"
        ],
        "summary": "One board, with its columns and labels.",
        "description": "The cards are a separate call, because a board with a thousand cards on it is not a response anybody wanted from a request for the column names.",
        "responses": {
          "200": {
            "description": "The board.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "operationId": "renameBoard",
        "tags": [
          "boards"
        ],
        "summary": "Change the name, the slug, or both.",
        "description": "A field left out keeps its current value. Changing the slug changes the URL of the board and of its page, and the old one stops working.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The board as it is now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Another board already has that slug.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteBoard",
        "tags": [
          "boards"
        ],
        "summary": "Delete a board and everything on it.",
        "description": "Columns, labels, cards, comments. There is no undo and no archive for a board.",
        "responses": {
          "204": {
            "description": "Gone."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/layout": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "put": {
        "operationId": "setLayout",
        "tags": [
          "boards"
        ],
        "summary": "Columns side by side, or stacked as rows.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Set."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/boards/{board}/columns": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "post": {
        "operationId": "createColumn",
        "tags": [
          "boards"
        ],
        "summary": "Add a column at the end.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new column.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Column"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/boards/{board}/columns/{column}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        },
        {
          "$ref": "#/components/parameters/column"
        }
      ],
      "patch": {
        "operationId": "updateColumn",
        "tags": [
          "boards"
        ],
        "summary": "Rename a column or change its WIP limit.",
        "description": "Absent fields keep the value they have, so a rename need not repeat the limit. Lowering the limit below the number of cards already in the column is allowed; the limit is checked when a card arrives.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnPatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteColumn",
        "tags": [
          "boards"
        ],
        "summary": "Delete a column.",
        "description": "The cards in it are deleted with it unless move_to names another column of the same board.",
        "parameters": [
          {
            "name": "move_to",
            "in": "query",
            "required": false,
            "description": "Column id on the same board that inherits the cards.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Gone."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/columns/{column}/cards/order": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        },
        {
          "$ref": "#/components/parameters/column"
        }
      ],
      "put": {
        "operationId": "reorderCards",
        "tags": [
          "cards"
        ],
        "summary": "Set the order of one column, moving cards into it as needed.",
        "description": "This is the drag and drop the board page does. A card named in the order that currently sits in another column of the board moves here, so a one-element list moves one card. A card of this column left out of the list keeps its place behind the ones named.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Ordered."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "The column is at its WIP limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/labels": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "post": {
        "operationId": "createLabel",
        "tags": [
          "boards"
        ],
        "summary": "Create a label on the board.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new label.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Label"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/labels/{label}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        },
        {
          "$ref": "#/components/parameters/label"
        }
      ],
      "patch": {
        "operationId": "updateLabel",
        "tags": [
          "boards"
        ],
        "summary": "Rename a label or recolour it.",
        "description": "Absent fields keep the value they have, so a rename need not repeat the colour.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelPatch"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteLabel",
        "tags": [
          "boards"
        ],
        "summary": "Delete a label.",
        "description": "It comes off every card that carries it.",
        "responses": {
          "204": {
            "description": "Gone."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/boards/{board}/cards": {
      "parameters": [
        {
          "$ref": "#/components/parameters/board"
        }
      ],
      "get": {
        "operationId": "listCards",
        "tags": [
          "cards"
        ],
        "summary": "The cards on the board, archived ones excluded.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "The same search the board page runs, syntax included. A bare word matches the title, the description or the name of a label the card carries, and every word has to match, so adding one narrows. The filters are label:<name>, which wants the name exactly; assignee:<part of the address> or assignee:none for unassigned; due: with one of none, overdue, today, week; and is:archived, which searches the archive instead of the board. An unrecognised prefix is searched for as text, because a colon in a title is ordinary: is:overdue finds cards that say that, not cards that are late. Use due:overdue for those.",
            "schema": {
              "type": "string"
            },
            "examples": {
              "text": {
                "value": "invoice"
              },
              "filters": {
                "value": "label:bug assignee:her@example.com due:overdue"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The cards, by column and then by position.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Card"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createCard",
        "tags": [
          "cards"
        ],
        "summary": "File a card.",
        "description": "column_id is optional and defaults to the first column of the board, so a script that files a card without caring where it lands does not have to read the board first.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new card.",
            "headers": {
              "Location": {
                "description": "Where the new card can be read.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "The column is at its WIP limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/cards/{card}/assignee": {
      "parameters": [
        {
          "$ref": "#/components/parameters/card"
        }
      ],
      "put": {
        "operationId": "setAssignee",
        "tags": [
          "cards"
        ],
        "summary": "Assign a card, or take it off everybody's list.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssigneeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The card as it is now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "The request is malformed, a field is unknown, or the service refused a value. field names the offending input where there is one.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Error": {
        "description": "Every failure has this body, including a 404 for an endpoint that does not exist and a 500.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
...
  }
}